+++ obj/src/convert/TypeConverter.d.ts
@@ -4,4 +4,5 @@ export declare class TypeConverter {
     static toNullableType<T>(type: TypeCode, value: any): T;
     static toType<T>(type: TypeCode, value: any): T;
     static toTypeWithDefault<T>(type: TypeCode, value: any, defaultValue: T): T;
+    static toString(type: TypeCode): string;
 }