+++ src/validate/FilterParamsSchema.ts
@@ -1,0 +1,10 @@
+import { TypeCode } from '../convert/TypeCode';
+import { MapSchema } from './MapSchema';
+
+export class FilterParamsSchema extends MapSchema {
+
+    public constructor() {
+        super(null, null, TypeCode.String, null);
+    }
+
+}